requirements.txt
langchain>=0.0.267
langchain-community>=0.0.10
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
tiktoken>=0.5.1
anthropic>=0.5.0
python-dotenv>=1.0.0